+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+### [144.2.2](https://github.com/Sage/carbon/compare/v144.2.1...v144.2.2) (2024-10-29)
+
+
+### Bug Fixes
+
+* inline label props have unexpected behaviour with new validation ([922694f](https://github.com/Sage/carbon/commit/922694f654e3c8cd7f9f49c56892be29828fb1fd))
+
+### [144.2.1](https://github.com/Sage/carbon/compare/v144.2.0...v144.2.1) (2024-10-28)
+
+
+### Bug Fixes
+
+* **multi-action-button, split-button:** ensure child button are accessible ([a3e2302](https://github.com/Sage/carbon/commit/a3e2302def7e97e32d1f28053706d79a2a14d5d9)), closes [#7005](https://github.com/Sage/carbon/issues/7005)
+
 ## [144.2.0](https://github.com/Sage/carbon/compare/v144.1.1...v144.2.0) (2024-10-25)